Cleaned up model, document and controllers API.
