+++ tests/controller/datacontroller.js
@@ -11,9 +11,6 @@ import buildViewConverter from '../../src/conversion/buildviewconverter';
 import buildModelConverter from '../../src/conversion/buildmodelconverter';
 
 import ModelDocumentFragment from '../../src/model/documentfragment';
-import ModelText from '../../src/model/text';
-import ModelRange from '../../src/model/range';
-import ModelSelection from '../../src/model/selection';
 
 import ViewDocumentFragment from '../../src/view/documentfragment';
 
@@ -334,225 +331,6 @@ describe( 'DataController', () => {
 			data.destroy();
 
 			expect( data ).to.respondTo( 'destroy' );
-		} );
-	} );
-
-	describe( 'insertContent()', () => {
-		it( 'should be decorated', () => {
-			schema.allow( { name: '$text', inside: '$root' } ); // To surpress warnings.
-
-			const spy = sinon.spy();
-
-			data.on( 'insertContent', spy );
-
-			data.insertContent( new ModelText( 'a' ), modelDocument.selection );
-
-			expect( spy.calledOnce ).to.be.true;
-		} );
-
-		it( 'should insert content (item)', () => {
-			schema.registerItem( 'paragraph', '$block' );
-
-			setData( model, '<paragraph>fo[]ar</paragraph>' );
-
-			data.insertContent( new ModelText( 'ob' ), modelDocument.selection );
-
-			expect( getData( model ) ).to.equal( '<paragraph>foob[]ar</paragraph>' );
-		} );
-
-		it( 'should insert content (document fragment)', () => {
-			schema.registerItem( 'paragraph', '$block' );
-
-			setData( model, '<paragraph>fo[]ar</paragraph>' );
-
-			data.insertContent( new ModelDocumentFragment( [ new ModelText( 'ob' ) ] ), modelDocument.selection );
-
-			expect( getData( model ) ).to.equal( '<paragraph>foob[]ar</paragraph>' );
-		} );
-
-		it( 'should use parent batch', () => {
-			schema.registerItem( 'paragraph', '$block' );
-			setData( model, '<paragraph>[]</paragraph>' );
-
-			model.change( writer => {
-				data.insertContent( new ModelText( 'abc' ), modelDocument.selection );
-				expect( writer.batch.deltas ).to.length( 1 );
-			} );
-		} );
-	} );
-
-	describe( 'deleteContent()', () => {
-		it( 'should be decorated', () => {
-			const spy = sinon.spy();
-
-			data.on( 'deleteContent', spy );
-
-			data.deleteContent( modelDocument.selection );
-
-			expect( spy.calledOnce ).to.be.true;
-		} );
-
-		it( 'should delete selected content', () => {
-			schema.registerItem( 'paragraph', '$block' );
-
-			setData( model, '<paragraph>fo[ob]ar</paragraph>' );
-
-			data.deleteContent( modelDocument.selection );
-
-			expect( getData( model ) ).to.equal( '<paragraph>fo[]ar</paragraph>' );
-		} );
-
-		it( 'should use parent batch', () => {
-			schema.registerItem( 'paragraph', '$block' );
-
-			setData( model, '<paragraph>fo[ob]ar</paragraph>' );
-
-			model.change( writer => {
-				data.deleteContent( modelDocument.selection );
-				expect( writer.batch.deltas ).to.length( 1 );
-			} );
-		} );
-	} );
-
-	describe( 'modifySelection()', () => {
-		it( 'should be decorated', () => {
-			schema.registerItem( 'paragraph', '$block' );
-			setData( model, '<paragraph>fo[ob]ar</paragraph>' );
-
-			const spy = sinon.spy();
-
-			data.on( 'modifySelection', spy );
-
-			data.modifySelection( modelDocument.selection );
-
-			expect( spy.calledOnce ).to.be.true;
-		} );
-
-		it( 'should modify a selection', () => {
-			schema.registerItem( 'paragraph', '$block' );
-
-			setData( model, '<paragraph>fo[ob]ar</paragraph>' );
-
-			data.modifySelection( modelDocument.selection, { direction: 'backward' } );
-
-			expect( getData( model ) ).to.equal( '<paragraph>fo[o]bar</paragraph>' );
-		} );
-	} );
-
-	describe( 'getSelectedContent()', () => {
-		it( 'should be decorated', () => {
-			const spy = sinon.spy();
-			const sel = new ModelSelection();
-
-			data.on( 'getSelectedContent', spy );
-
-			data.getSelectedContent( sel );
-
-			expect( spy.calledOnce ).to.be.true;
-		} );
-
-		it( 'should return selected content', () => {
-			schema.registerItem( 'paragraph', '$block' );
-
-			setData( model, '<paragraph>fo[ob]ar</paragraph>' );
-
-			const content = data.getSelectedContent( modelDocument.selection );
-
-			expect( stringify( content ) ).to.equal( 'ob' );
-		} );
-
-		it( 'should use parent batch', () => {
-			schema.registerItem( 'paragraph', '$block' );
-
-			setData( model, '<paragraph>fo[ob]ar</paragraph>' );
-
-			model.change( writer => {
-				data.getSelectedContent( modelDocument.selection );
-				expect( writer.batch.deltas ).to.length( 1 );
-			} );
-		} );
-	} );
-
-	describe( 'hasContent()', () => {
-		let root;
-
-		beforeEach( () => {
-			schema.registerItem( 'paragraph', '$block' );
-			schema.registerItem( 'div', '$block' );
-			schema.allow( { name: '$block', inside: 'div' } );
-			schema.registerItem( 'image' );
-			schema.allow( { name: 'image', inside: 'div' } );
-			schema.objects.add( 'image' );
-
-			setData(
-				model,
-
-				'<div>' +
-					'<paragraph></paragraph>' +
-				'</div>' +
-				'<paragraph>foo</paragraph>' +
-				'<div>' +
-					'<image></image>' +
-				'</div>'
-			);
-
-			root = modelDocument.getRoot();
-		} );
-
-		it( 'should return true if given element has text node', () => {
-			const pFoo = root.getChild( 1 );
-
-			expect( data.hasContent( pFoo ) ).to.be.true;
-		} );
-
-		it( 'should return true if given element has element that is an object', () => {
-			const divImg = root.getChild( 2 );
-
-			expect( data.hasContent( divImg ) ).to.be.true;
-		} );
-
-		it( 'should return false if given element has no elements', () => {
-			const pEmpty = root.getChild( 0 ).getChild( 0 );
-
-			expect( data.hasContent( pEmpty ) ).to.be.false;
-		} );
-
-		it( 'should return false if given element has only elements that are not objects', () => {
-			const divP = root.getChild( 0 );
-
-			expect( data.hasContent( divP ) ).to.be.false;
-		} );
-
-		it( 'should return true if there is a text node in given range', () => {
-			const range = ModelRange.createFromParentsAndOffsets( root, 1, root, 2 );
-
-			expect( data.hasContent( range ) ).to.be.true;
-		} );
-
-		it( 'should return true if there is a part of text node in given range', () => {
-			const pFoo = root.getChild( 1 );
-			const range = ModelRange.createFromParentsAndOffsets( pFoo, 1, pFoo, 2 );
-
-			expect( data.hasContent( range ) ).to.be.true;
-		} );
-
-		it( 'should return true if there is element that is an object in given range', () => {
-			const divImg = root.getChild( 2 );
-			const range = ModelRange.createFromParentsAndOffsets( divImg, 0, divImg, 1 );
-
-			expect( data.hasContent( range ) ).to.be.true;
-		} );
-
-		it( 'should return false if range is collapsed', () => {
-			const range = ModelRange.createFromParentsAndOffsets( root, 1, root, 1 );
-
-			expect( data.hasContent( range ) ).to.be.false;
-		} );
-
-		it( 'should return false if range has only elements that are not objects', () => {
-			const range = ModelRange.createFromParentsAndOffsets( root, 0, root, 1 );
-
-			expect( data.hasContent( range ) ).to.be.false;
 		} );
 	} );
 } );
