+++ tests/model/utils/getselectedcontent.js
@@ -3,27 +3,25 @@
  * For licensing, see LICENSE.md.
  */
 
-import Model from '../../src/model/model';
-import DataController from '../../src/controller/datacontroller';
-import DocumentFragment from '../../src/model/documentfragment';
-import getSelectedContent from '../../src/controller/getselectedcontent';
-import { setData, stringify } from '../../src/dev-utils/model';
+import Model from '../../../src/model/model';
+import DocumentFragment from '../../../src/model/documentfragment';
+import getSelectedContent from '../../../src/model/utils/getselectedcontent';
+import { setData, stringify } from '../../../src/dev-utils/model';
 
 describe( 'DataController utils', () => {
-	let model, doc, data;
+	let model, doc;
 
 	describe( 'getSelectedContent', () => {
 		it( 'should use parent batch', () => {
 			model = new Model();
 			doc = model.document;
 			doc.createRoot();
-			data = new DataController( model );
 
 			model.schema.allow( { name: '$text', inside: '$root' } );
 			setData( model, 'x[abc]x' );
 
 			model.change( writer => {
-				getSelectedContent( data, doc.selection );
+				getSelectedContent( model, doc.selection );
 				expect( writer.batch.deltas ).to.length( 1 );
 			} );
 		} );
@@ -33,7 +31,6 @@ describe( 'DataController utils', () => {
 				model = new Model();
 				doc = model.document;
 				doc.createRoot();
-				data = new DataController( model );
 
 				const schema = model.schema;
 
@@ -48,7 +45,7 @@ describe( 'DataController utils', () => {
 			it( 'returns empty fragment for no selection', () => {
 				setData( model, 'abc' );
 
-				const frag = getSelectedContent( data, doc.selection );
+				const frag = getSelectedContent( model, doc.selection );
 
 				expect( frag ).instanceOf( DocumentFragment );
 				expect( frag.isEmpty ).to.be.true;
@@ -57,7 +54,7 @@ describe( 'DataController utils', () => {
 			it( 'returns empty fragment for empty selection', () => {
 				setData( model, 'a[]bc' );
 
-				const frag = getSelectedContent( data, doc.selection );
+				const frag = getSelectedContent( model, doc.selection );
 
 				expect( frag ).instanceOf( DocumentFragment );
 				expect( frag.isEmpty ).to.be.true;
@@ -66,7 +63,7 @@ describe( 'DataController utils', () => {
 			it( 'gets one character', () => {
 				setData( model, 'a[b]c' );
 
-				const frag = getSelectedContent( data, doc.selection );
+				const frag = getSelectedContent( model, doc.selection );
 				const content = stringify( frag );
 
 				expect( frag ).instanceOf( DocumentFragment );
@@ -76,49 +73,49 @@ describe( 'DataController utils', () => {
 			it( 'gets full text', () => {
 				setData( model, '[abc]' );
 
-				const content = stringify( getSelectedContent( data, doc.selection ) );
+				const content = stringify( getSelectedContent( model, doc.selection ) );
 				expect( content ).to.equal( 'abc' );
 			} );
 
 			it( 'gets text with an attribute', () => {
 				setData( model, 'xxx<$text bold="true">a[b]c</$text>' );
 
-				const content = stringify( getSelectedContent( data, doc.selection ) );
+				const content = stringify( getSelectedContent( model, doc.selection ) );
 				expect( content ).to.equal( '<$text bold="true">b</$text>' );
 			} );
 
 			it( 'gets text with attributes', () => {
 				setData( model, 'x<$text bold="true">a[b</$text><$text italic="true">c]d</$text>x' );
 
-				const content = stringify( getSelectedContent( data, doc.selection ) );
+				const content = stringify( getSelectedContent( model, doc.selection ) );
 				expect( content ).to.equal( '<$text bold="true">b</$text><$text italic="true">c</$text>' );
 			} );
 
 			it( 'gets text with and without attribute', () => {
 				setData( model, '<$text bold="true">a[b</$text>c]d' );
 
-				const content = stringify( getSelectedContent( data, doc.selection ) );
+				const content = stringify( getSelectedContent( model, doc.selection ) );
 				expect( content ).to.equal( '<$text bold="true">b</$text>c' );
 			} );
 
 			it( 'gets text and element', () => {
 				setData( model, '[ab<image></image>c]' );
 
-				const content = stringify( getSelectedContent( data, doc.selection ) );
+				const content = stringify( getSelectedContent( model, doc.selection ) );
 				expect( content ).to.equal( 'ab<image></image>c' );
 			} );
 
 			it( 'gets one element', () => {
 				setData( model, 'a[<image></image>]b' );
 
-				const content = stringify( getSelectedContent( data, doc.selection ) );
+				const content = stringify( getSelectedContent( model, doc.selection ) );
 				expect( content ).to.equal( '<image></image>' );
 			} );
 
 			it( 'gets multiple elements', () => {
 				setData( model, '[<image></image><image></image>]' );
 
-				const content = stringify( getSelectedContent( data, doc.selection ) );
+				const content = stringify( getSelectedContent( model, doc.selection ) );
 				expect( content ).to.equal( '<image></image><image></image>' );
 			} );
 		} );
@@ -128,7 +125,6 @@ describe( 'DataController utils', () => {
 				model = new Model();
 				doc = model.document;
 				doc.createRoot();
-				data = new DataController( model );
 
 				const schema = model.schema;
 
@@ -148,63 +144,63 @@ describe( 'DataController utils', () => {
 			it( 'gets one character', () => {
 				setData( model, '<paragraph>a[b]c</paragraph>' );
 
-				const content = stringify( getSelectedContent( data, doc.selection ) );
+				const content = stringify( getSelectedContent( model, doc.selection ) );
 				expect( content ).to.equal( 'b' );
 			} );
 
 			it( 'gets entire paragraph content', () => {
 				setData( model, '<paragraph>[a<image></image>b]</paragraph>' );
 
-				const content = stringify( getSelectedContent( data, doc.selection ) );
+				const content = stringify( getSelectedContent( model, doc.selection ) );
 				expect( content ).to.equal( 'a<image></image>b' );
 			} );
 
 			it( 'gets two blocks - partial, partial', () => {
 				setData( model, '<heading1>a[bc</heading1><paragraph>de]f</paragraph>' );
 
-				const content = stringify( getSelectedContent( data, doc.selection ) );
+				const content = stringify( getSelectedContent( model, doc.selection ) );
 				expect( content ).to.equal( '<heading1>bc</heading1><paragraph>de</paragraph>' );
 			} );
 
 			it( 'gets two blocks - full, partial', () => {
 				setData( model, '<heading1>[abc</heading1><paragraph>de]f</paragraph>' );
 
-				const content = stringify( getSelectedContent( data, doc.selection ) );
+				const content = stringify( getSelectedContent( model, doc.selection ) );
 				expect( content ).to.equal( '<heading1>abc</heading1><paragraph>de</paragraph>' );
 			} );
 
 			it( 'gets two blocks - full, partial 2', () => {
 				setData( model, '<heading1>[abc</heading1><paragraph>de<image></image>]f</paragraph>' );
 
-				const content = stringify( getSelectedContent( data, doc.selection ) );
+				const content = stringify( getSelectedContent( model, doc.selection ) );
 				expect( content ).to.equal( '<heading1>abc</heading1><paragraph>de<image></image></paragraph>' );
 			} );
 
 			it( 'gets two blocks - full, partial 3', () => {
 				setData( model, '<heading1>x</heading1><heading1>[abc</heading1><paragraph><image></image>de]f</paragraph>' );
 
-				const content = stringify( getSelectedContent( data, doc.selection ) );
+				const content = stringify( getSelectedContent( model, doc.selection ) );
 				expect( content ).to.equal( '<heading1>abc</heading1><paragraph><image></image>de</paragraph>' );
 			} );
 
 			it( 'gets two blocks - full, partial 4', () => {
 				setData( model, '<heading1>[abc</heading1><paragraph>de]f<image></image></paragraph>' );
 
-				const content = stringify( getSelectedContent( data, doc.selection ) );
+				const content = stringify( getSelectedContent( model, doc.selection ) );
 				expect( content ).to.equal( '<heading1>abc</heading1><paragraph>de</paragraph>' );
 			} );
 
 			it( 'gets two blocks - partial, full', () => {
 				setData( model, '<heading1>a[bc</heading1><paragraph>def]</paragraph>' );
 
-				const content = stringify( getSelectedContent( data, doc.selection ) );
+				const content = stringify( getSelectedContent( model, doc.selection ) );
 				expect( content ).to.equal( '<heading1>bc</heading1><paragraph>def</paragraph>' );
 			} );
 
 			it( 'gets two blocks - partial, full 2', () => {
 				setData( model, '<heading1>a[<image></image>bc</heading1><paragraph>def]</paragraph>' );
 
-				const content = stringify( getSelectedContent( data, doc.selection ) );
+				const content = stringify( getSelectedContent( model, doc.selection ) );
 				expect( content ).to.equal( '<heading1><image></image>bc</heading1><paragraph>def</paragraph>' );
 			} );
 
@@ -212,7 +208,7 @@ describe( 'DataController utils', () => {
 			it( 'gets two blocks - empty, full', () => {
 				setData( model, '<heading1>abc[</heading1><paragraph>def]</paragraph>' );
 
-				const content = stringify( getSelectedContent( data, doc.selection ) );
+				const content = stringify( getSelectedContent( model, doc.selection ) );
 				expect( content ).to.equal( '<paragraph>def</paragraph>' );
 			} );
 
@@ -220,28 +216,28 @@ describe( 'DataController utils', () => {
 			it( 'gets two blocks - partial, empty', () => {
 				setData( model, '<heading1>a[bc</heading1><paragraph>]def</paragraph>' );
 
-				const content = stringify( getSelectedContent( data, doc.selection ) );
+				const content = stringify( getSelectedContent( model, doc.selection ) );
 				expect( content ).to.equal( '<heading1>bc</heading1>' );
 			} );
 
 			it( 'gets three blocks', () => {
 				setData( model, '<heading1>a[bc</heading1><paragraph>x</paragraph><paragraph>de]f</paragraph>' );
 
-				const content = stringify( getSelectedContent( data, doc.selection ) );
+				const content = stringify( getSelectedContent( model, doc.selection ) );
 				expect( content ).to.equal( '<heading1>bc</heading1><paragraph>x</paragraph><paragraph>de</paragraph>' );
 			} );
 
 			it( 'gets block image', () => {
 				setData( model, '<paragraph>a</paragraph>[<blockImage><caption>Foo</caption></blockImage>]<paragraph>b</paragraph>' );
 
-				const content = stringify( getSelectedContent( data, doc.selection ) );
+				const content = stringify( getSelectedContent( model, doc.selection ) );
 				expect( content ).to.equal( '<blockImage><caption>Foo</caption></blockImage>' );
 			} );
 
 			it( 'gets two blocks', () => {
 				setData( model, '<paragraph>a</paragraph>[<blockImage></blockImage><blockImage></blockImage>]<paragraph>b</paragraph>' );
 
-				const content = stringify( getSelectedContent( data, doc.selection ) );
+				const content = stringify( getSelectedContent( model, doc.selection ) );
 				expect( content ).to.equal( '<blockImage></blockImage><blockImage></blockImage>' );
 			} );
 
@@ -249,7 +245,7 @@ describe( 'DataController utils', () => {
 			it( 'gets content when multiple text items needs to be removed from the right excess', () => {
 				setData( model, '<paragraph>a[b</paragraph><paragraph>c]d<$text bold="true">e</$text>f</paragraph>' );
 
-				const content = stringify( getSelectedContent( data, doc.selection ) );
+				const content = stringify( getSelectedContent( model, doc.selection ) );
 				expect( content )
 					.to.equal( '<paragraph>b</paragraph><paragraph>c</paragraph>' );
 			} );
@@ -258,7 +254,7 @@ describe( 'DataController utils', () => {
 			it( 'gets content when multiple text items needs to be removed from the left excess', () => {
 				setData( model, '<paragraph>a<$text bold="true">b</$text>c[d</paragraph><paragraph>e]f</paragraph>' );
 
-				const content = stringify( getSelectedContent( data, doc.selection ) );
+				const content = stringify( getSelectedContent( model, doc.selection ) );
 				expect( content )
 					.to.equal( '<paragraph>d</paragraph><paragraph>e</paragraph>' );
 			} );
@@ -269,7 +265,6 @@ describe( 'DataController utils', () => {
 				model = new Model();
 				doc = model.document;
 				doc.createRoot();
-				data = new DataController( model );
 
 				const schema = model.schema;
 
@@ -284,28 +279,28 @@ describe( 'DataController utils', () => {
 			it( 'gets content when ends are equally deeply nested', () => {
 				setData( model, '<heading1>x</heading1><quote><paragraph>a[bc</paragraph><paragraph>de]f</paragraph></quote>' );
 
-				const content = stringify( getSelectedContent( data, doc.selection ) );
+				const content = stringify( getSelectedContent( model, doc.selection ) );
 				expect( content ).to.equal( '<paragraph>bc</paragraph><paragraph>de</paragraph>' );
 			} );
 
 			it( 'gets content when left end nested deeper', () => {
 				setData( model, '<quote><paragraph>a[bc</paragraph></quote><paragraph>de]f</paragraph>' );
 
-				const content = stringify( getSelectedContent( data, doc.selection ) );
+				const content = stringify( getSelectedContent( model, doc.selection ) );
 				expect( content ).to.equal( '<quote><paragraph>bc</paragraph></quote><paragraph>de</paragraph>' );
 			} );
 
 			it( 'gets content when left end nested deeper 2', () => {
 				setData( model, '<quote><paragraph>a[bc</paragraph><heading1>x</heading1></quote><paragraph>de]f</paragraph>' );
 
-				const content = stringify( getSelectedContent( data, doc.selection ) );
+				const content = stringify( getSelectedContent( model, doc.selection ) );
 				expect( content ).to.equal( '<quote><paragraph>bc</paragraph><heading1>x</heading1></quote><paragraph>de</paragraph>' );
 			} );
 
 			it( 'gets content when left end nested deeper 3', () => {
 				setData( model, '<quote><heading1>x</heading1><paragraph>a[bc</paragraph></quote><paragraph>de]f</paragraph>' );
 
-				const content = stringify( getSelectedContent( data, doc.selection ) );
+				const content = stringify( getSelectedContent( model, doc.selection ) );
 				expect( content ).to.equal( '<quote><paragraph>bc</paragraph></quote><paragraph>de</paragraph>' );
 			} );
 
@@ -313,21 +308,21 @@ describe( 'DataController utils', () => {
 			it( 'gets content when left end nested deeper 4', () => {
 				setData( model, '<quote><heading1>x[</heading1><paragraph>abc</paragraph></quote><paragraph>de]f</paragraph>' );
 
-				const content = stringify( getSelectedContent( data, doc.selection ) );
+				const content = stringify( getSelectedContent( model, doc.selection ) );
 				expect( content ).to.equal( '<quote><paragraph>abc</paragraph></quote><paragraph>de</paragraph>' );
 			} );
 
 			it( 'gets content when right end nested deeper', () => {
 				setData( model, '<paragraph>a[bc</paragraph><quote><paragraph>de]f</paragraph></quote>' );
 
-				const content = stringify( getSelectedContent( data, doc.selection ) );
+				const content = stringify( getSelectedContent( model, doc.selection ) );
 				expect( content ).to.equal( '<paragraph>bc</paragraph><quote><paragraph>de</paragraph></quote>' );
 			} );
 
 			it( 'gets content when both ends nested deeper than the middle element', () => {
 				setData( model, '<quote><heading1>a[bc</heading1></quote><heading1>x</heading1><quote><heading1>de]f</heading1></quote>' );
 
-				const content = stringify( getSelectedContent( data, doc.selection ) );
+				const content = stringify( getSelectedContent( model, doc.selection ) );
 				expect( content )
 					.to.equal( '<quote><heading1>bc</heading1></quote><heading1>x</heading1><quote><heading1>de</heading1></quote>' );
 			} );
@@ -347,7 +342,7 @@ describe( 'DataController utils', () => {
 					'</quote>'
 				);
 
-				const content = stringify( getSelectedContent( data, doc.selection ) );
+				const content = stringify( getSelectedContent( model, doc.selection ) );
 				expect( content )
 					.to.equal( '<paragraph>ar</paragraph>bo' );
 			} );
